+++ QoL/SkipIntro/mod_info.js
@@ -23,8 +23,10 @@
 
    /* Config files will be created when user click "Reset". */
    DefaultConfig : {
+      "Config_Version" : 20200322,
       "Skip_Logos" : true,
       "Skip_HottestYear" : true,
+      "Skip_NewGameIntro" : false,
       "Skip_Landings" : true,
       "Skip_CurtainDrop" : true,
       "Skip_CurtainLift" : true,